Update README wit rudimentary installation instructions.  Tweak freshdb.sh to handle case in which sqlite file doesn't not exist.

+++ db/freshdb.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
 
-rm drafts.sqlite
+if [ -e drafts.sqlite ] 
+then
+  rm drafts.sqlite
+fi
 sqlite3 drafts.sqlite < db.schema
